+++ app/src/main/java/com/smashing/app/data/mapper/user/GetUserInfoDetailMapper.kt
@@ -3,9 +3,7 @@ package com.smashing.app.data.mapper.user
 import com.smashing.app.data.model.profile.ProfileInfo
 import com.smashing.app.data.model.profile.ProfileItem
 import com.smashing.app.data.model.profile.user.UserProfileInfo
-import com.smashing.app.data.model.review.GameReviewResult
 import com.smashing.app.data.remote.dto.user.GetUserInfoDetailResponse
-import com.smashing.app.data.remote.dto.user.GetUserRecentReviewStatsResponse
 import com.smashing.app.data.type.GenderType
 import com.smashing.app.data.type.SportType
 import com.smashing.app.data.type.TierType
@@ -23,7 +21,7 @@ fun GetUserInfoDetailResponse.toUserProfileInfo(): UserProfileInfo =
     )
 
 
-fun GetUserInfoDetailResponse.SelectedProfile.toSelectedProfile(): ProfileInfo {
+private fun GetUserInfoDetailResponse.SelectedProfile.toSelectedProfile(): ProfileInfo {
     return ProfileInfo(
         profileId = profileId,
         sportType = SportType.findSportTypeToSportCode(sportCode),
@@ -36,21 +34,9 @@ fun GetUserInfoDetailResponse.SelectedProfile.toSelectedProfile(): ProfileInfo {
     )
 }
 
-fun GetUserInfoDetailResponse.Profile.toUserProfileItem(): ProfileItem =
+private fun GetUserInfoDetailResponse.Profile.toUserProfileItem(): ProfileItem =
     ProfileItem(
         profileId = profileId,
         sportType = SportType.findSportTypeToSportCode(this.sportCode),
         isActive = isSelected,
     )
-
-fun GetUserRecentReviewStatsResponse.toGameReviewResult(): GameReviewResult =
-    GameReviewResult(
-        bestCount = this.ratingCounts.best,
-        goodCount = this.ratingCounts.good,
-        badCount = this.ratingCounts.bad,
-        goodMannerCount = this.tagCounts.goodManner,
-        onTimeCount = this.tagCounts.onTime,
-        fairPlayCount = this.tagCounts.fairPlay,
-        fastResponseCount = this.tagCounts.fastResponse,
-    )
-
